Sprint 5
Functionality to delete a project, education or employment record has been implemented. To be tested via unit testing but looks to work correctly. User only has the ability to delete records that belong to that user. Error message appears if the user attempts to delete a record that does not associate itself to the employee number of the user that is authenticated into the system (logged in).